@@ -1,4 +1,5 @@
-﻿using HarSA.Startups;
+﻿using Autofac;
+using HarSA.Startups;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.Configuration;
@@ -14,6 +15,11 @@ namespace HarSA.AspNetCore.Api.Infrastructure
         public virtual void Configure(IApplicationBuilder application)
         {
             application.UseAuthentication();
+        }
+
+        public virtual void ConfigureContainer(ContainerBuilder containerBuilder, IConfiguration configuration)
+        {
+
         }
 
         public virtual void ConfigureServices(IServiceCollection services, IConfiguration configuration)
