@@ -1,18 +1,27 @@
 ﻿using Autofac;
 using HarSA.AspNetCore.Mvc.Notifications;
-using HarSA.Configurations;
-using HarSA.Dependency;
+using HarSA.Startups;
+using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace HarSA.AspNetCore.Mvc.Infrastructure
 {
-    public class NotificationRegistrar : IDependencyRegistrar
+    public class NotificationRegistrar : IAppStartup
     {
         public int Order => 9000;
 
-        public void Register(ContainerBuilder builder, ITypeFinder typeFinder, HarConfig config, IConfiguration configuration)
+        public void Configure(IApplicationBuilder application)
         {
-            builder.RegisterType<NotificationService>().As<INotificationService>().InstancePerLifetimeScope();
+        }
+
+        public void ConfigureContainer(ContainerBuilder containerBuilder, IConfiguration configuration)
+        {
+            containerBuilder.RegisterType<NotificationService>().As<INotificationService>().InstancePerLifetimeScope();
+        }
+
+        public void ConfigureServices(IServiceCollection services, IConfiguration configuration)
+        {
         }
     }
 }
