@@ -1,23 +1,32 @@
 ﻿using Autofac;
-using HarSA.Configurations;
-using HarSA.Dependency;
 using HarSA.EntityFrameworkCore.Application;
 using HarSA.EntityFrameworkCore.Repositories;
+using HarSA.Startups;
+using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace HarSA.EntityFrameworkCore.Infrastructure
 {
-    public class EntityFrameworkCoreRegistrar : IDependencyRegistrar
+    public class EntityFrameworkCoreRegistrar : IAppStartup
     {
         public int Order => 100;
 
-        public void Register(ContainerBuilder builder, ITypeFinder typeFinder, HarConfig config, IConfiguration configuration)
+        public void Configure(IApplicationBuilder application)
         {
-            builder.RegisterGeneric(typeof(BaseRepo<>)).As(typeof(IRepo<>)).InstancePerLifetimeScope();
-            builder.RegisterGeneric(typeof(CrudService<>)).As(typeof(ICrudService<>)).InstancePerLifetimeScope();
+        }
 
-            builder.RegisterGeneric(typeof(EfRepository<>)).As(typeof(IRepository<>)).InstancePerLifetimeScope();
-            builder.RegisterGeneric(typeof(HarCrudService<>)).As(typeof(IHarCrudService<>)).InstancePerLifetimeScope();
+        public void ConfigureContainer(ContainerBuilder containerBuilder, IConfiguration configuration)
+        {
+            containerBuilder.RegisterGeneric(typeof(BaseRepo<>)).As(typeof(IRepo<>)).InstancePerLifetimeScope();
+            containerBuilder.RegisterGeneric(typeof(CrudService<>)).As(typeof(ICrudService<>)).InstancePerLifetimeScope();
+
+            containerBuilder.RegisterGeneric(typeof(EfRepository<>)).As(typeof(IRepository<>)).InstancePerLifetimeScope();
+            containerBuilder.RegisterGeneric(typeof(HarCrudService<>)).As(typeof(IHarCrudService<>)).InstancePerLifetimeScope();
+        }
+
+        public void ConfigureServices(IServiceCollection services, IConfiguration configuration)
+        {
         }
     }
 }
