@@ -1,11 +1,9 @@
 ﻿using Autofac;
 using Autofac.Extensions.DependencyInjection;
 using HarSA.Configurations;
-using HarSA.Dependency;
 using HarSA.Exceptions;
 using HarSA.Startups;
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using System;
@@ -16,68 +14,25 @@ namespace HarSA
 {
     public class HarEngine : IEngine
     {
-        public IServiceProvider ServiceProvider { get; private set; }
+        public ILifetimeScope AutofacContainer { get; set; }
 
-        protected IServiceProvider GetServiceProvider()
-        {
-            var accessor = ServiceProvider.GetRequiredService<IHttpContextAccessor>();
-            var context = accessor.HttpContext;
-
-            return context != null ? context.RequestServices : ServiceProvider;
-        }
-
-        protected virtual IServiceProvider RegisterDependencies(IServiceCollection services, ITypeFinder typeFinder, HarConfig config, IConfiguration configuration)
-        {
-            var containerBuilder = new ContainerBuilder();
-
-            // register engine
-            // services.AddSingleton<IEngine>(this);
-            containerBuilder.RegisterInstance(this).As<IEngine>().SingleInstance();
-            // register typeFinder
-            // services.AddSingleton(typeFinder);
-            containerBuilder.RegisterInstance(typeFinder).As<ITypeFinder>().SingleInstance();
-
-            //populate Autofac container builder with the set of registered service descriptors
-            containerBuilder.Populate(services);
-
-            var dependencyRegistrars = typeFinder.FindClassesOfType<IDependencyRegistrar>();
-
-            var instances = dependencyRegistrars.Select(registrar => (IDependencyRegistrar)Activator.CreateInstance(registrar)).OrderBy(registrar => registrar.Order);
-
-            foreach (var instance in instances)
-            {
-                instance.Register(containerBuilder, typeFinder, config, configuration);
-            }
-
-            // create service provider
-            // _serviceProvider = services.BuildServiceProvider();
-            ServiceProvider = new AutofacServiceProvider(containerBuilder.Build());
-            return ServiceProvider;
-        }
-
-        public IServiceProvider ConfigureEngineServices(IServiceCollection services, IConfiguration configuration, HarConfig config)
+        public void ConfigureEngineServices(IServiceCollection services, IConfiguration configuration, HarConfig config)
         {
             var typeFinder = new AppDomainTypeFinder();
             var startupConfigurations = typeFinder.FindClassesOfType<IAppStartup>();
-
-            var instances = startupConfigurations.Select(startup => (IAppStartup)Activator.CreateInstance(startup)).OrderBy(startup => startup.Order);
-
+            var instances = startupConfigurations.Select(t => (IAppStartup)Activator.CreateInstance(t)).OrderBy(o => o.Order);
             foreach (var instance in instances)
             {
                 instance.ConfigureServices(services, configuration);
             }
-
-            RegisterDependencies(services, typeFinder, config, configuration);
-
-            return ServiceProvider;
         }
 
         public void ConfigureEngineRequestPipeline(IApplicationBuilder application)
         {
+            AutofacContainer = application.ApplicationServices.GetAutofacRoot();
             var typeFinder = Resolve<ITypeFinder>();
             var startupConfigurations = typeFinder.FindClassesOfType<IAppStartup>();
-
-            var instances = startupConfigurations.Select(startup => (IAppStartup)Activator.CreateInstance(startup)).OrderBy(startup => startup.Order);
+            var instances = startupConfigurations.Select(t => (IAppStartup)Activator.CreateInstance(t)).OrderBy(o => o.Order);
 
             foreach (var instance in instances)
             {
@@ -85,28 +40,40 @@ namespace HarSA
             }
         }
 
+        public void ConfigureContainer(ContainerBuilder builder, IConfiguration configuration)
+        {
+            var typeFinder = new AppDomainTypeFinder();
+
+            builder.RegisterInstance(this).As<IEngine>().SingleInstance();
+
+            builder.RegisterInstance(typeFinder).As<ITypeFinder>().SingleInstance();
+            var startupConfigurations = typeFinder.FindClassesOfType<IAppStartup>();
+            var instances = startupConfigurations.Select(t => (IAppStartup)Activator.CreateInstance(t)).OrderBy(o => o.Order);
+            foreach (var instance in instances)
+            {
+                instance.ConfigureContainer(builder, configuration);
+            }
+        }
+
         public T Resolve<T>() where T : class
         {
-            return (T)GetServiceProvider().GetRequiredService(typeof(T));
+            return (T)Resolve(typeof(T));
         }
 
         public object Resolve(Type type)
         {
-            return GetServiceProvider().GetRequiredService(type);
+            return AutofacContainer.Resolve(type);
         }
 
         public IEnumerable<T> ResolveAll<T>()
         {
-            return (IEnumerable<T>)GetServiceProvider().GetServices(typeof(T));
+            return AutofacContainer.Resolve<IEnumerable<T>>();
         }
 
         public object ResolveUnregistered(Type type)
         {
             Exception innerException = null;
-
-            var constructors = type.GetConstructors();
-
-            foreach (var constructor in constructors)
+            foreach (var constructor in type.GetConstructors())
             {
                 try
                 {
@@ -127,6 +94,7 @@ namespace HarSA
                     innerException = ex;
                 }
             }
+
             throw new HarException("No constructor was found that had all the dependencies satisfied.", innerException);
         }
     }
