@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Builder;
+﻿using Autofac;
+using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -7,6 +8,8 @@ namespace HarSA.Startups
     public interface IAppStartup
     {
         void ConfigureServices(IServiceCollection services, IConfiguration configuration);
+
+        void ConfigureContainer(ContainerBuilder containerBuilder, IConfiguration configuration);
 
         void Configure(IApplicationBuilder application);
 
